@@ -1,6 +1,7 @@
 {
   "name": "@amermathsoc/mathjax-unicode-math",
   "main": "./js/unicode-math.js",
+  "type": "module",
   "version": "1.2.3",
   "description": "MathJax TeX input extension reimplementing (parts of) the unicode-math package",
   "keywords": [
@@ -29,15 +30,14 @@
     "build": "npx tsc &&  node ./node_modules/mathjax-full/components/bin/makeAll"
   },
   "devDependencies": {
-    "@babel/core": "7.16.12",
-    "@babel/preset-env": "7.16.11",
-    "babel-loader": "8.2.3",
     "jsdom": "19.0.0",
-    "mathjax-full": "github:amermathsoc/MathJax-src#f85e13d343bd05df15270a723bfd605b5c8e65b9",
-    "standard-version": "9.3.2",
-    "terser-webpack-plugin": "5.3.0",
-    "typescript": "4.5.5",
-    "webpack": "5.67.0",
-    "webpack-cli": "^4.9.1"
+    "mathjax-full": "4.0.0-beta.3",
+    "mathjax-stix2-font": "4.0.0-beta.3",
+    "standard-version": "9.5.0",
+    "terser-webpack-plugin": "5.3.9",
+    "typescript": "^5.1.3",
+    "typescript-tools": "^0.3.1",
+    "webpack": "^5.85.0",
+    "webpack-cli": "^5.1.1"
   }
 }
